# opt: add rule to fold JSON field access into Values operator