--- a/pkg/sql/opt/norm/project_funcs.go
+++ b/pkg/sql/opt/norm/project_funcs.go
@@ -17,6 +17,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
+	"github.com/cockroachdb/cockroach/pkg/util/json"
 	"github.com/cockroachdb/errors"
 )
 
@@ -98,43 +99,29 @@ func (c *CustomFuncs) MergeProjectWithValues(
 	})
 }
 
-// CanUnnestTuplesFromValues returns true if the Values operator has a single
-// column containing tuples that can be unfolded into multiple columns.
-//
+// CanUnnestTuplesFromValues returns true if the given single-column Values
+// operator has tuples that can be unfolded into multiple columns.
 // This is the case if:
-//
-// 	1. The Values operator has exactly one output column.
-//
-// 	2. The single output column is of type tuple.
-//
-// 	3. There is at least one row.
-//
-// 	4. All tuples in the single column are either TupleExpr's or ConstExpr's
+// 	1. The single output column is of type tuple.
+// 	2. All tuples in the single column are either TupleExpr's or ConstExpr's
 //     that wrap DTuples, as opposed to dynamically generated tuples.
-//
 func (c *CustomFuncs) CanUnnestTuplesFromValues(expr memo.RelExpr) bool {
 	values := expr.(*memo.ValuesExpr)
-	if !c.HasOneCol(expr) {
-		return false
-	}
 	colTypeFam := c.mem.Metadata().ColumnMeta(values.Cols[0]).Type.Family()
 	if colTypeFam != types.TupleFamily {
 		return false
 	}
-	if len(values.Rows) < 1 {
-		return false
-	}
-	for _, row := range values.Rows {
-		if !c.IsStaticTuple(row.(*memo.TupleExpr).Elems[0]) {
+	for i := range values.Rows {
+		if !c.IsStaticTuple(values.Rows[i].(*memo.TupleExpr).Elems[0]) {
 			return false
 		}
 	}
 	return true
 }
 
-// OnlyTupleColumnsAccessed ensures that the input ProjectionsExpr contains no
+// HasNoDirectTupleReferences ensures that the input ProjectionsExpr contains no
 // direct references to the tuple represented by the given ColumnID.
-func (c *CustomFuncs) OnlyTupleColumnsAccessed(
+func (c *CustomFuncs) HasNoDirectTupleReferences(
 	projections memo.ProjectionsExpr, tupleCol opt.ColumnID,
 ) bool {
 	var check func(expr opt.Expr) bool
@@ -217,8 +204,8 @@ func (c *CustomFuncs) UnnestTuplesFromValues(
 		case *memo.ConstExpr:
 			dTuple := t.Value.(*tree.DTuple)
 			tupleVals := make(memo.ScalarListExpr, len(dTuple.D))
-			for i, v := range dTuple.D {
-				val := c.f.ConstructConstVal(v, tupleType.TupleContents()[i])
+			for i := range dTuple.D {
+				val := c.f.ConstructConstVal(dTuple.D[i], tupleType.TupleContents()[i])
 				tupleVals[i] = val
 			}
 			outTuples[i] = c.f.ConstructTuple(tupleVals, tupleType)
@@ -263,6 +250,279 @@ func (c *CustomFuncs) FoldTupleColumnAccess(
 			replace(projection.Element).(opt.ScalarExpr), projection.Col)
 	}
 	return newProjections
+}
+
+// CanUnnestJSONFromValues returns true if the Values operator has a single
+// column containing JSON expressions that can be unfolded into multiple
+// columns. This is the case if:
+//  1. The projections don't directly reference the JSON itself. Only fields
+//     within the JSON object can be referenced.
+//  2. All JSON keys referenced by the projections are present in the first row.
+//  3. All JSON keys present in the first row are present in all other rows.
+// CanUnnestJSONFromValues should only be called if the Values operator has a
+// single column and at least one row.
+//
+// Note: technically we only need to check that the JSON fields referenced by
+// the projections exist in all the rows. For simplicity, we instead check that
+// (1) all references exist in the first row and (2) that all keys from the
+// first row also exist in all other rows.
+func (c *CustomFuncs) CanUnnestJSONFromValues(
+	relExpr memo.RelExpr, projections memo.ProjectionsExpr, jsonCol opt.ColumnID,
+) bool {
+	values := relExpr.(*memo.ValuesExpr)
+	colTypeFam := c.mem.Metadata().ColumnMeta(values.Cols[0]).Type.Family()
+	if colTypeFam != types.JsonFamily {
+		return false
+	}
+	// Retrieve the JSON expression from the first row.
+	constExpr, ok := values.Rows[0].(*memo.TupleExpr).Elems[0].(*memo.ConstExpr)
+	if !ok {
+		// The contents of the JSON expression can't be statically determined. (This
+		// can happen when the row contains a reference to a JSON table column).
+		return false
+	}
+	firstJSON, ok := constExpr.Value.(*tree.DJSON)
+	if !ok {
+		return false
+	}
+	if !(firstJSON.Type() == json.ObjectJSONType) {
+		// The JSON expression must be a JSON object with key-value pairs referenced
+		// by the projections.
+		return false
+	}
+
+	// Recursively traverses an expression, looking for direct references to
+	// jsonCol. Also ensures that any references to JSON fields are part of the
+	// first row's schema. For example, this query is valid:
+	//
+	//    SELECT j->'x'
+	//    FROM
+	//    (VALUES
+	//        ('{"x": "one"}'::JSON),
+	//        ('{"x": "two", "y": 2}'::JSON)
+	//    ) v(j);
+	//
+	// But these queries are not valid:
+	//
+	//    SELECT j
+	//    FROM
+	//    (VALUES
+	//        ('{"x": "one"}'::JSON),
+	//        ('{"x": "two", "y": 2}'::JSON)
+	//    ) v(j);
+	//
+	//    SELECT j->'x', j->'y'
+	//    FROM
+	//    (VALUES
+	//        ('{"x": "one"}'::JSON),
+	//        ('{"x": "two", "y": 2}'::JSON)
+	//    ) v(j);
+	//
+	var check func(expr opt.Expr) bool
+	check = func(expr opt.Expr) bool {
+		switch t := expr.(type) {
+		case *memo.FetchValExpr:
+			if v, ok := t.Json.(*memo.VariableExpr); ok {
+				if constExpr, ok := t.Index.(*memo.ConstExpr); ok {
+					if key, ok := constExpr.Value.(*tree.DString); ok {
+						if v.Col == jsonCol {
+							// Ensure that the key this projection is referencing exists in
+							// the first row.
+							exists, err := firstJSON.Exists(string(*key))
+							return exists && err == nil
+						}
+						return true
+					}
+				}
+			}
+			return false
+
+		case *memo.VariableExpr:
+			// Ensure that the JSON column itself is not referenced; only its fields
+			// can be referenced.
+			return t.Col != jsonCol
+		}
+		for i, n := 0, expr.ChildCount(); i < n; i++ {
+			if !check(expr.Child(i)) {
+				return false
+			}
+		}
+		return true
+	}
+
+	// Traverse all ProjectionsItems.
+	for i := range projections {
+		if !check(projections[i].Element) {
+			return false
+		}
+	}
+
+	// Ensure that all Values rows are ConstExpr's that wrap DJSON datums. Also
+	// ensure that the DJSON datums have at least the keys from the first row.
+	// This check is performed after the projections are walked because there may
+	// be many Values rows.
+	for i := 1; i < len(values.Rows); i++ {
+		expr := values.Rows[i].(*memo.TupleExpr).Elems[0]
+		if !c.IsConstJSON(expr) {
+			// The contents of this JSON expression cannot be statically determined.
+			return false
+		}
+		currJSON := expr.(*memo.ConstExpr).Value.(*tree.DJSON)
+		iter, err := firstJSON.ObjectIter()
+		if err != nil {
+			return false
+		}
+		// Iterate through the keys of the first JSON object and ensure that they
+		// all exist in every other row. For example, these are a valid set of rows:
+		//
+		//    ('{"x": 1}'::JSON), ('{"x": 2, "y": 'two'}'::JSON)
+		//
+		// but these are not:
+		//
+		//    ('{"x": 2, "y": 'two'}'::JSON), ('{"x": 1}'::JSON)
+		//
+		for iter.Next() {
+			if exists, err := currJSON.Exists(iter.Key()); !exists || err != nil {
+				return false
+			}
+		}
+	}
+
+	return true
+}
+
+// UnnestJSONFromValues takes in a Values operator that has a single column
+// of JSON expressions and a ColList with columns corresponding to each JSON key
+// in the first row of the ValuesExpr. It returns a new Values operator with the
+// json fields expanded out into the Values rows.
+func (c *CustomFuncs) UnnestJSONFromValues(expr memo.RelExpr, newCols opt.ColList) memo.RelExpr {
+	values := expr.(*memo.ValuesExpr)
+	outTuples := make(memo.ScalarListExpr, len(values.Rows))
+	valTypes := make([]*types.T, len(newCols))
+	for i := range valTypes {
+		valTypes[i] = types.Jsonb
+	}
+	tupleType := types.MakeTuple(valTypes)
+
+	// Construct new rows with values corresponding to the keys from first row.
+	// As an example, given Values rows like the following:
+	//
+	//    ('{"x": 1, "y": 'one'}'::JSON), ('{"x": 2, "y": 'two', "z": 'zwei'}'::JSON)
+	//
+	// The output rows will be:
+	//
+	//    (1::JSON, 'one'::JSON), (2::JSON, 'two'::JSON)
+	//
+	// Any values corresponding to keys that are not present in the first row will
+	// be discarded.
+	firstJSON := values.Rows[0].(*memo.TupleExpr).Elems[0].(*memo.ConstExpr).Value.(*tree.DJSON)
+	for i := range values.Rows {
+		rowExpr := values.Rows[i].(*memo.TupleExpr).Elems[0]
+		tupleVals := make(memo.ScalarListExpr, len(newCols))
+		iter, err := firstJSON.ObjectIter()
+		if err != nil {
+			panic(errors.AssertionFailedf("failed to retrieve ObjectIter: %v", err))
+		}
+		// Iterate through the keys of the first row and use them to get the values
+		// from the current row. Add these values to the new Values rows.
+		idx := 0
+		for iter.Next() {
+			jsonVal, err := rowExpr.(*memo.ConstExpr).Value.(*tree.DJSON).FetchValKey(iter.Key())
+			if err != nil {
+				panic(errors.AssertionFailedf("FetchValKey failed: %v", err))
+			}
+			dJSON := tree.NewDJSON(jsonVal)
+			tupleVals[idx] = c.f.ConstructConstVal(dJSON, types.Jsonb)
+			idx++
+		}
+		outTuples[i] = c.f.ConstructTuple(tupleVals, tupleType)
+	}
+
+	// Construct ValuesExpr with the new rows.
+	valuesPrivate := &memo.ValuesPrivate{Cols: newCols, ID: c.mem.Metadata().NextUniqueID()}
+	return c.f.ConstructValues(outTuples, valuesPrivate)
+}
+
+// FoldJSONFieldAccess constructs a new ProjectionsExpr from the old one with
+// any FetchVal operators that refer to the original JSON column (oldColID)
+// replaced by new Variables wrapping columns from the output of the given
+// ValuesExpr. Example:
+//
+//    SELECT j->'a' AS j_a FROM ...
+// =>
+//    SELECT j_a FROM ...
+//
+func (c *CustomFuncs) FoldJSONFieldAccess(
+	projections memo.ProjectionsExpr,
+	newCols opt.ColList,
+	oldColID opt.ColumnID,
+	oldInput memo.RelExpr,
+) memo.ProjectionsExpr {
+	newProjections := make(memo.ProjectionsExpr, len(projections))
+	oldValues := oldInput.(*memo.ValuesExpr)
+
+	// Create a mapping from JSON keys to the new columns that were created for
+	// them.
+	keysToCols := make(map[string]opt.ColumnID)
+	firstJSON := oldValues.Rows[0].(*memo.TupleExpr).Elems[0].(*memo.ConstExpr).Value.(*tree.DJSON)
+	iter, err := firstJSON.ObjectIter()
+	if err != nil {
+		panic(errors.AssertionFailedf("failed to retrieve ObjectIter: %v", err))
+	}
+	idx := 0
+	for iter.Next() {
+		keysToCols[iter.Key()] = newCols[idx]
+		idx++
+	}
+
+	// Recursively traverses a ProjectionsItem element and replaces references to
+	// json fields with one of the newly constructed columns.
+	var replace ReplaceFunc
+	replace = func(nd opt.Expr) opt.Expr {
+		if fetchVal, ok := nd.(*memo.FetchValExpr); ok {
+			if variable, ok := fetchVal.Json.(*memo.VariableExpr); ok {
+				// Skip past references to columns other than the input tuple column.
+				if variable.Col == oldColID {
+					// Replace this reference to a JSON field with the column that was
+					// created to replace it.
+					key := string(*fetchVal.Index.(*memo.ConstExpr).Value.(*tree.DString))
+					return c.f.ConstructVariable(keysToCols[key])
+				}
+			}
+		}
+		return c.f.Replace(nd, replace)
+	}
+
+	// Construct and return a new ProjectionsExpr using the new ColumnIDs.
+	for i := range projections {
+		projection := &projections[i]
+		newProjections[i] = c.f.ConstructProjectionsItem(
+			replace(projection.Element).(opt.ScalarExpr), projection.Col)
+	}
+	return newProjections
+}
+
+// MakeColsForUnnestJSON creates a new column for each key in the first row of
+// the given ValuesExpr, and returns a ColList with those columns. The columns
+// will be in the same order as the corresponding keys from the first row.
+func (c *CustomFuncs) MakeColsForUnnestJSON(input memo.RelExpr, jsonCol opt.ColumnID) opt.ColList {
+	values := input.(*memo.ValuesExpr)
+	dJSON := values.Rows[0].(*memo.TupleExpr).Elems[0].(*memo.ConstExpr).Value.(*tree.DJSON)
+	mem := c.mem.Metadata()
+	jsonAlias := mem.ColumnMeta(jsonCol).Alias
+
+	// Create a new column for each JSON key found in dJSON.
+	iter, err := dJSON.ObjectIter()
+	if err != nil {
+		panic(errors.AssertionFailedf("failed to retrieve ObjectIter: %v", err))
+	}
+	newColIDs := make(opt.ColList, 0, dJSON.Len())
+	for iter.Next() {
+		newAlias := fmt.Sprintf("%s_%s", jsonAlias, iter.Key())
+		newColID := mem.AddColumn(newAlias, types.Jsonb)
+		newColIDs = append(newColIDs, newColID)
+	}
+	return newColIDs
 }
 
 // CanPushColumnRemappingIntoValues returns true if there is at least one