--- a/pkg/sql/opt/xform/testdata/external/trading
+++ b/pkg/sql/opt/xform/testdata/external/trading
@@ -1262,13 +1262,13 @@ upsert transactiondetails
  ├── cardinality: [0 - 0]
  ├── stable+volatile, side-effects, mutations
  ├── project
- │    ├── columns: upsert_dealerid:31 upsert_isbuy:32 upsert_transactiondate:33 upsert_cardid:34 sellprice:29 buyprice:30 "?column?":11!null bool:12!null current_timestamp:13 int8:14 int8:15 column18:18 transactiondetails.dealerid:21 transactiondetails.isbuy:22 transactiondate:23 cardid:24 quantity:25 transactiondetails.sellprice:26 transactiondetails.buyprice:27 transactiondetails.version:28
+ │    ├── columns: upsert_dealerid:31 upsert_isbuy:32 upsert_transactiondate:33 upsert_cardid:34 sellprice:29 buyprice:30 "?column?":11!null bool:12!null current_timestamp:13 int8:14!null int8:15!null column18:18 transactiondetails.dealerid:21 transactiondetails.isbuy:22 transactiondate:23 cardid:24 quantity:25 transactiondetails.sellprice:26 transactiondetails.buyprice:27 transactiondetails.version:28
  │    ├── cardinality: [1 - 2]
  │    ├── stable+volatile, side-effects
  │    ├── lax-key: (13-15,21-25)
  │    ├── fd: ()-->(11,12), (13-15)~~>(18), (21-25)-->(26-28), (21)-->(31), (21,22)-->(32), (13,21,23)-->(33), (14,21,24)-->(34), (13-15,21-25)~~>(18,29,30)
  │    ├── left-join (lookup transactiondetails)
- │    │    ├── columns: "?column?":11!null bool:12!null current_timestamp:13 int8:14 int8:15 column18:18 sellprice:19 buyprice:20 transactiondetails.dealerid:21 transactiondetails.isbuy:22 transactiondate:23 cardid:24 quantity:25 transactiondetails.sellprice:26 transactiondetails.buyprice:27 transactiondetails.version:28
+ │    │    ├── columns: "?column?":11!null bool:12!null current_timestamp:13 int8:14!null int8:15!null column18:18 sellprice:19 buyprice:20 transactiondetails.dealerid:21 transactiondetails.isbuy:22 transactiondate:23 cardid:24 quantity:25 transactiondetails.sellprice:26 transactiondetails.buyprice:27 transactiondetails.version:28
  │    │    ├── key columns: [11 12 13 14 15] = [21 22 23 24 25]
  │    │    ├── lookup columns are key
  │    │    ├── cardinality: [1 - 2]
@@ -1276,32 +1276,32 @@ upsert transactiondetails
  │    │    ├── lax-key: (13-15,21-25)
  │    │    ├── fd: ()-->(11,12), (13-15)~~>(18-20), (21-25)-->(26-28)
  │    │    ├── ensure-upsert-distinct-on
- │    │    │    ├── columns: "?column?":11!null bool:12!null current_timestamp:13 int8:14 int8:15 column18:18 sellprice:19 buyprice:20
- │    │    │    ├── grouping columns: current_timestamp:13 int8:14 int8:15
+ │    │    │    ├── columns: "?column?":11!null bool:12!null current_timestamp:13 int8:14!null int8:15!null column18:18 sellprice:19 buyprice:20
+ │    │    │    ├── grouping columns: current_timestamp:13 int8:14!null int8:15!null
  │    │    │    ├── error: "UPSERT or INSERT...ON CONFLICT command cannot affect row a second time"
  │    │    │    ├── cardinality: [1 - 2]
  │    │    │    ├── stable+volatile, side-effects
  │    │    │    ├── lax-key: (13-15)
  │    │    │    ├── fd: ()-->(11,12), (13-15)~~>(11,12,18-20)
  │    │    │    ├── project
- │    │    │    │    ├── columns: sellprice:19 buyprice:20 column18:18 "?column?":11!null bool:12!null current_timestamp:13 int8:14 int8:15
+ │    │    │    │    ├── columns: sellprice:19 buyprice:20 column18:18 "?column?":11!null bool:12!null current_timestamp:13 int8:14!null int8:15!null
  │    │    │    │    ├── cardinality: [2 - 2]
  │    │    │    │    ├── stable+volatile, side-effects
  │    │    │    │    ├── fd: ()-->(11,12)
  │    │    │    │    ├── values
- │    │    │    │    │    ├── columns: detail:10!null
+ │    │    │    │    │    ├── columns: detail_b:54!null detail_c:55!null detail_q:56!null detail_s:57!null
  │    │    │    │    │    ├── cardinality: [2 - 2]
- │    │    │    │    │    ├── ('{"b": 2.29, "c": 49833, "q": 4, "s": 2.89}',)
- │    │    │    │    │    └── ('{"b": 17.59, "c": 29483, "q": 2, "s": 18.93}',)
+ │    │    │    │    │    ├── ('2.29', '49833', '4', '2.89')
+ │    │    │    │    │    └── ('17.59', '29483', '2', '18.93')
  │    │    │    │    └── projections
- │    │    │    │         ├── crdb_internal.round_decimal_values((detail:10->'s')::STRING::DECIMAL(10,4), 4) [as=sellprice:19, outer=(10), immutable]
- │    │    │    │         ├── crdb_internal.round_decimal_values((detail:10->'b')::STRING::DECIMAL(10,4), 4) [as=buyprice:20, outer=(10), immutable]
+ │    │    │    │         ├── crdb_internal.round_decimal_values(detail_s:57::STRING::DECIMAL(10,4), 4) [as=sellprice:19, outer=(57), immutable]
+ │    │    │    │         ├── crdb_internal.round_decimal_values(detail_b:54::STRING::DECIMAL(10,4), 4) [as=buyprice:20, outer=(54), immutable]
  │    │    │    │         ├── cluster_logical_timestamp() [as=column18:18, volatile, side-effects]
  │    │    │    │         ├── 1 [as="?column?":11]
  │    │    │    │         ├── false [as=bool:12]
  │    │    │    │         ├── current_timestamp() [as=current_timestamp:13, stable, side-effects]
- │    │    │    │         ├── (detail:10->'c')::STRING::INT8 [as=int8:14, outer=(10), immutable]
- │    │    │    │         └── (detail:10->'q')::STRING::INT8 [as=int8:15, outer=(10), immutable]
+ │    │    │    │         ├── detail_c:55::STRING::INT8 [as=int8:14, outer=(55), immutable]
+ │    │    │    │         └── detail_q:56::STRING::INT8 [as=int8:15, outer=(56), immutable]
  │    │    │    └── aggregations
  │    │    │         ├── first-agg [as=sellprice:19, outer=(19)]
  │    │    │         │    └── sellprice:19