… pkg/sql/opt/norm/testdata/rules/project
… pkg/sql/opt/xform/testdata/external/trading-mutation