Previously, the optimizer didn't fold JSON field access over an input
Values operator.

This patch adds a rule that constructs a new Values operator with columns
corresponding to each referenced field in the JSON expression. It replaces
the surrounding Project operator with a new one that references the new
columns and takes the new Values as input. Example:
```
SELECT cust->'id' AS id, cust->'name' AS name
FROM (VALUES
   ('{"id": 1, "name": 'Drew'}'::JSON),
   ('{"id": 2, "name": 'Radu'}'::JSON),
   ('{"id": 3, "name": 'Rebecca'}'::JSON)
) v(cust)
=>
SELECT id, name
FROM (VALUES
   (1::JSON, 'Drew'::JSON),
   (2::JSON, 'Radu'::JSON),
   (3::JSON, 'Rebecca'::JSON)
) v(id, name)
```
Release note: None