@DrewKimball 2020-06-12
I benchmarked this query with different numbers of rows:

```
SELECT j->'a', j->'b', j->'c', j->'d', j->'e'
FROM (VALUES
	('{"a": 1, "b": 2, "c": 3, "d": 4, "e": 5}'::JSON),
         …
);
```
**1 Row:**
```
name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd    76.2µs ± 3%    76.4µs ± 3%   ~     (p=0.436 n=10+10)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd    18.3kB ± 0%    18.3kB ± 0%   ~     (p=0.857 n=8+8)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd       150 ± 0%       150 ± 0%   ~     (all equal)
```
**10 Rows:**
```
name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     126µs ± 1%     111µs ± 2%  -11.85%  (p=0.000 n=9+10)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd    35.1kB ± 0%    32.8kB ± 0%   -6.61%  (p=0.000 n=8+8)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd       557 ± 0%       466 ± 0%  -16.34%  (p=0.000 n=8+8)
```
**100 Rows:**
```
name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     458µs ± 3%     418µs ± 3%  -8.67%  (p=0.000 n=10+10)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     160kB ± 1%     166kB ± 0%  +3.49%  (p=0.000 n=9+9)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     3.81k ± 0%     3.62k ± 0%  -4.82%  (p=0.000 n=10+8)
```
**1,000 Rows:**
```
name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd    3.35ms ± 2%    3.26ms ± 2%  -2.74%  (p=0.000 n=10+9)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd    1.39MB ± 0%    1.50MB ± 0%  +7.73%  (p=0.000 n=9+8)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     36.3k ± 0%     35.2k ± 0%  -2.98%  (p=0.000 n=9+8)
```
**10,000 Rows:**
```
name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd    34.8ms ± 3%    33.1ms ± 1%  -5.01%  (p=0.000 n=10+8)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd    13.7MB ± 0%    14.9MB ± 0%  +8.12%  (p=0.000 n=8+8)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd      361k ± 0%      351k ± 0%  -2.79%  (p=0.000 n=8+8)
```
**100,000 Rows:**
```
name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     398ms ± 3%     353ms ± 5%  -11.23%  (p=0.000 n=10+10)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     137MB ± 0%     148MB ± 0%   +8.14%  (p=0.000 n=9+8)

name                              old time/op    new time/op    delta
EndToEnd/json-fold-test/EndToEnd     3.61M ± 0%     3.51M ± 0%   -2.78%  (p=0.000 n=9+8)
```