@rytaft 2020-06-15
___
*[pkg/sql/opt/norm/general_funcs.go, line 176 at r2](https://reviewable.io/reviews/cockroachdb/cockroach/50131#-M9i8Uj12HBBA6PLEIkh:-M9tTgLh-Oior4e0-91S:bo5wa54) ([raw file](https://github.com/cockroachdb/cockroach/blob/9f222fe92f02e80ff883979aa590a161166bda91/pkg/sql/opt/norm/general_funcs.go#L176)):*
<details><summary><i>Previously, DrewKimball (Drew Kimball) wrote…</i></summary><blockquote>

Done.
</blockquote></details>

looks like you didn't push the latest update....


<!-- Sent from Reviewable.io -->